Store the groups registry per versioned group
This lets us avoid using a weakref dictionary to have a global reference.

Unfortunately, this does not appear to fix the garbage collection bug with
hdf5 (#162). That seems to indicate that the issue is not directly related to
weakrefs. Either way, this does seem like it might be a better design for the
registry of groups.